Order of Curve
The curve used in this code has an equation x^3+x+2 mod 11.  The algorithm determines the order of the elliptic curve when a generator point is given. In this example, we fixed the generator point to (4,5). The algorithm continues to increment n until the point at infinity is reached. Point at infinity is a resultant point of addition of two points on elliptic curve where x1=x2 and y1=-y2